ctci ch1 review

+++ LeetCode/SetMatrixZero.java
@@ -2,7 +2,8 @@
     Given a m x n matrix, if an element is 0, set its entire row and column to 0. Do it in place.
 */
 
-// use first row and col to store status of matrix
+// Check first whether the first row and first column should be set to zeros
+// then use first row and col as extra storage
 // time: O(m*n); space: O(1)
 public class Solution {
     public void setZeroes(int[][] matrix) {
@@ -24,7 +25,7 @@ public class Solution {
                 }
             }
         }
-        for (int i = 1; i < h; i++){
+        for (int i = 1; i < h; i++){        // Note these 2 for loops can quickly set zeros
             for (int j = 1; j < w; j++){
                 if (matrix[0][j] == 0 || matrix[i][0] == 0){
                     matrix[i][j] = 0;
@@ -42,12 +43,13 @@ public class Solution {
     }
 }
 
-// use bit vector, max size of matrix should be no bigger than 32
+// use bit vector, max size of matrix should be no bigger than 64
 // time: O(m*n); space:O(1)
+// Note: this solution cannot pass OJ
 public class Solution {
     public void setZeroes(int[][] matrix) {
-        int row = 0;
-        int col = 0;
+        long row = 0;
+        long col = 0;
         for (int i = 0; i < matrix.length; i++) {
             for (int j = 0; j < matrix[0].length; j++) {
                 if (matrix[i][j] == 0) {
